requirements.txt
opencv-python
numpy
pillow
pyautogui
Wand
scikit-image
